refactor (rename table) : table -> hall
