+++ src/main/java/kitchenpos/hall/bo/TableBo.java
@@ -1,9 +1,9 @@
-package kitchenpos.table.bo;
+package kitchenpos.hall.bo;
 
 import kitchenpos.order.dao.OrderDao;
 import kitchenpos.order.dao.OrderTableDao;
 import kitchenpos.order.model.OrderStatus;
-import kitchenpos.table.model.OrderTable;
+import kitchenpos.hall.model.OrderTable;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
