+++ src/main/java/kitchenpos/hall/model/TableGroup.java
@@ -1,4 +1,4 @@
-package kitchenpos.table.model;
+package kitchenpos.hall.model;
 
 import java.time.LocalDateTime;
 import java.util.List;
