+++ src/main/java/kitchenpos/order/bo/OrderBo.java
@@ -7,7 +7,7 @@ import kitchenpos.order.dao.OrderTableDao;
 import kitchenpos.order.model.Order;
 import kitchenpos.order.model.OrderLineItem;
 import kitchenpos.order.model.OrderStatus;
-import kitchenpos.table.model.OrderTable;
+import kitchenpos.hall.model.OrderTable;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.util.CollectionUtils;
