+++ src/test/java/kitchenpos/hall/TableBoTest.java
@@ -1,4 +1,4 @@
-package kitchenpos.table;
+package kitchenpos.hall;
 
 import kitchenpos.order.InMemoryOrderDao;
 import kitchenpos.order.InMemoryOrderTableDao;
@@ -6,8 +6,8 @@ import kitchenpos.order.dao.OrderDao;
 import kitchenpos.order.dao.OrderTableDao;
 import kitchenpos.order.model.Order;
 import kitchenpos.order.model.OrderStatus;
-import kitchenpos.table.model.OrderTable;
-import kitchenpos.table.bo.TableBo;
+import kitchenpos.hall.model.OrderTable;
+import kitchenpos.hall.bo.TableBo;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
