+++ src/test/java/kitchenpos/hall/TableGroupBoTest.java
@@ -1,12 +1,12 @@
-package kitchenpos.table;
+package kitchenpos.hall;
 
 import kitchenpos.order.InMemoryOrderDao;
 import kitchenpos.order.InMemoryOrderTableDao;
 import kitchenpos.order.dao.OrderDao;
 import kitchenpos.order.dao.OrderTableDao;
-import kitchenpos.table.dao.TableGroupDao;
-import kitchenpos.table.model.TableGroup;
-import kitchenpos.table.bo.TableGroupBo;
+import kitchenpos.hall.dao.TableGroupDao;
+import kitchenpos.hall.model.TableGroup;
+import kitchenpos.hall.bo.TableGroupBo;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
