+++ src/test/java/kitchenpos/hall/controller/TableGroupRestControllerTest.java
@@ -1,7 +1,7 @@
-package kitchenpos.table.controller;
+package kitchenpos.hall.controller;
 
-import kitchenpos.table.bo.TableGroupBo;
-import kitchenpos.table.model.TableGroup;
+import kitchenpos.hall.bo.TableGroupBo;
+import kitchenpos.hall.model.TableGroup;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.autoconfigure.web.servlet.HttpEncodingAutoConfiguration;
